requirements.txt
Package                Version
---------------------- ----------
bcrypt                 5.0.0
bidict                 0.23.1
blinker                1.9.0
certifi                2025.11.12
charset-normalizer     3.4.4
click                  8.3.1
Flask                  3.0.0
Flask-Cors             4.0.0
Flask-SocketIO         5.3.6
h11                    0.16.0
idna                   3.11
itsdangerous           2.2.0
Jinja2                 3.1.6
MarkupSafe             3.0.3
mysql-connector-python 8.0.32
pip                    23.0.1
protobuf               3.20.3
PyMySQL                1.1.2
python-dotenv          1.0.0
python-engineio        4.9.0
python-socketio        5.11.0
requests               2.31.0
setuptools             65.5.0
simple-websocket       1.1.0
urllib3                2.5.0
Werkzeug               3.1.4
wheel                  0.45.1
wsproto                1.3.2
